requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
pandas==1.3.3
openpyxl==3.0.9
Pillow==8.3.2
numpy==1.21.2
Werkzeug==2.0.1